explicitly zero memory of the original buffer when reallocing
zero the defunct buffer after copying the contents to the
newly-allocated buffer when reallocing in the resize method.

This change (or one very similar to it) was made in the original
upstream pijul repository, which seems to have disappeared recently.
I have reproduced it from memory as my own commit for the purposes of
fixing the bug.